requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
facenet-pytorch>=2.5.0
Pillow>=10.0.0
mtcnn>=0.1.1
faiss-cpu>=1.7.4
flask>=3.0.0
pandas>=2.0.0
plotly>=5.0.0
eventlet>=0.33.0
Django>=4.2.0
